simplified batch searching, going to just get all in batch sessions
